requirements.txt
tensorflow
pillow
tensorflowjs